dropanalyzer-backend/requirements.txt
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask==2.3.3
PyJWT==2.8.0
aiohttp
alembic
bandit
celery[redis]
gunicorn
passlib>=1.7.4
psycopg2-binary
python-dotenv
redis
safety
semgrep
sqlalchemy
